# Optimize `colormap`