--- a/src/colormaps.jl
+++ b/src/colormaps.jl
@@ -122,7 +122,8 @@ distinguishable_colors(n::Integer; kwargs...) = distinguishable_colors(n, Vector
 """
     sequential_palette(h, N::Int=100; <keyword arguments>)
 
-Implements the color palette creation technique by [Wijffelaars, M., et al. (2008)](http://magnaview.nl/documents/MagnaView-M_Wijffelaars-Generating_color_palettes_using_intuitive_parameters.pdf).
+Implements the color palette creation technique by
+[Wijffelaars, M., et al. (2008)](https://dl.acm.org/doi/10.1111/j.1467-8659.2008.01203.x).
 
 Colormaps are formed using Bézier curves in LCHuv colorspace
 with some constant hue. In addition, start and end points can be given
@@ -130,137 +131,141 @@ that are then blended to the original hue smoothly.
 
 # Arguments
 
-- N        - number of colors
-- h        - the main hue [0,360]
-- c        - the overall lightness contrast [0,1]
-- s        - saturation [0,1]
-- b        - brightness [0,1]
-- w        - cold/warm parameter, i.e. the strength of the starting color [0,1]
-- d        - depth of the ending color [0,1]
-- wcolor   - starting color (warmness)
-- dcolor   - ending color (depth)
-- logscale - true/false for toggling logspacing
+- `N`        - number of colors
+- `h`        - the main hue [0,360]
+- `c`        - the overall lightness contrast [0,1]
+- `s`        - saturation [0,1]
+- `b`        - brightness [0,1]
+- `w`        - cold/warm parameter, i.e. the strength of the starting color [0,1]
+- `d`        - depth of the ending color [0,1]
+- `wcolor`   - starting color (warmness)
+- `dcolor`   - ending color (depth)
+- `logscale` - `true`/`false` for toggling logspacing
 """
 function sequential_palette(h,
                             N::Int=100;
+                            w=0.15,
+                            d=0.0,
                             c=0.88,
                             s=0.6,
                             b=0.75,
-                            w=0.15,
-                            d=0.0,
-                            wcolor=RGB(1,1,0),
-                            dcolor=RGB(0,0,1),
-                            logscale=false)
+                            wcolor=RGB(1.0, 1.0, 0.0),
+                            dcolor=RGB(0.0, 0.0, 1.0),
+                            logscale::Bool=false)
+    F = Float64
+    return _sequential_palette(N, logscale, F(h), F(w), F(d), F(c), F(s), F(b),
+                               RGB{F}(wcolor), RGB{F}(dcolor))
+end
 
-    function mix_hue(a, h0, h1)
+function _sequential_palette(N, logscale, h, w, d, c, s, b, wcolor, dcolor)
+    function mix_hue(h0, h1, w)
         m = normalize_hue(180.0 + h1 - h0) - 180.0
-        normalize_hue(h0 + a * m)
+        normalize_hue(h0 + w * m)
     end
-    function mix_linearly(a::C, b::C, s) where C<:Color
-        base_color_type(C)((1-s)*comp1(a)+s*comp1(b), (1-s)*comp2(a)+s*comp2(b), (1-s)*comp3(a)+s*comp3(b))
-    end
+    mix_linearly(a::Float64, b::Float64, w) = (1.0 - w) * a + w * b
+    mix_linearly(a::LCHuv{Float64}, b::LCHuv{Float64}, w) = weighted_color_mean(w, b, a)
 
-    pstart=convert(LCHuv, wcolor)
-    p1=MSC(h)
+    pstart = convert(LCHuv{Float64}, wcolor)
+    pend   = convert(LCHuv{Float64}, dcolor)
+    p1 = MSC(h)
     #p0=LCHuv(0,0,h) #original end point
-    pend=convert(LCHuv, dcolor)
 
     #multi-hue start point
-    p2l = 100 * (1. - w) + w * pstart.l
-    p2h=mix_hue(w,h,pstart.h)
-    p2c=min(MSC(p2h,p2l), w*s*pstart.c)
-    p2=LCHuv(p2l,p2c,p2h)
+    p2l = mix_linearly(100.0, pstart.l, w)
+    p2h = mix_hue(h, pstart.h, w)
+    p2c = min(MSC(p2h, p2l), w * s * pstart.c)
+    p2 = LCHuv{Float64}(p2l, p2c, p2h)
 
     #multi-hue ending point
-    p0l=20.0*d
-    p0h=mix_hue(d,h,pend.h)
-    p0c=min(MSC(p0h,p0l), d*s*pend.c)
-    p0=LCHuv(p0l,p0c,p0h)
+    p0l = 20.0 * d
+    p0h = mix_hue(h, pend.h, d)
+    p0c = min(MSC(p0h, p0l), d * s * pend.c)
+    p0 = LCHuv{Float64}(p0l, p0c, p0h)
 
-    q0=mix_linearly(p0,p1,s)
-    q2=mix_linearly(p2,p1,s)
-    q1=mix_linearly(q0,q2,0.5)
+    q0 = mix_linearly(p0, p1, s)
+    q2 = mix_linearly(p2, p1, s)
+    q1 = mix_linearly(q0, q2, 0.5)
 
-    pal = RGB{Float64}[]
+    pal = Vector{LCHuv{Float64}}(undef, N)
 
     if logscale
         absc = exp10.(range(-2., stop=0., length=N))
     else
-        absc = range(0.,stop=1.,length=N)
+        absc = range(0., stop=1., length=N)
     end
 
-    for t in absc
-        u=1.0-t
+    for (i, t) in enumerate(absc)
+        u = 1.0 - t
 
         #Change grid to favor light colors and to be uniform along the curve
-        u = (125.0-125.0*0.2^((1.0-c)*b+u*c))
-        u = invBezier(u,p0.l,p2.l,q0.l,q1.l,q2.l)
+        lt = 125.0 - 125.0 * 0.2^mix_linearly(b, u, c)
+        tt = inv_bezier(lt, p0.l, p2.l, q0.l, q1.l, q2.l)
 
         #Get color components from Bezier curves
-        ll = Bezier(u, p0.l, p2.l, q0.l, q1.l, q2.l)
-        cc = Bezier(u, p0.c, p2.c, q0.c, q1.c, q2.c)
-        hh = Bezier(u, p0.h, p2.h, q0.h, q1.h, q2.h)
+        ll = bezier(tt, p0.l, p2.l, q0.l, q1.l, q2.l)
+        cc = bezier(tt, p0.c, p2.c, q0.c, q1.c, q2.c)
+        hh = bezier(tt, p0.h, p2.h, q0.h, q1.h, q2.h)
 
-        push!(pal, convert(RGB, LCHuv(ll,cc,hh)))
+        @inbounds pal[i] =LCHuv{Float64}(ll, cc, hh)
     end
 
-    pal
+    RGB{Float64}.(pal)
 end
 
 """
     diverging_palette(h1, h2, N::Int=100; <keyword arguments>)
 
-Create diverging palettes by combining 2 sequential palettes
+Create diverging palettes by combining 2 sequential palettes.
 
 # Arguments
 
-- N        - number of colors
-- h1       - the main hue of the left side [0,360]
-- h2       - the main hue of the right side [0,360]
-- c        - the overall lightness contrast [0,1]
-- s        - saturation [0,1]
-- b        - brightness [0,1]
-- w        - cold/warm parameter, i.e. the strength of the starting color [0,1]
-- d1       - depth of the ending color in the left side [0,1]
-- d2       - depth of the ending color in the right side [0,1]
-- wcolor   - starting color (warmness)
-- dcolor1  - ending color of the left side (depth)
-- dcolor2  - ending color of the right side (depth)
-- logscale - true/false for toggling logspacing
+- `N`        - number of colors
+- `mid`      - the position of the midpoint (0,1)
+- `h1`       - the main hue of the first part [0,360]
+- `h2`       - the main hue of the latter part [0,360]
+- `c`        - the overall lightness contrast [0,1]
+- `s`        - saturation [0,1]
+- `b`        - brightness [0,1]
+- `w`        - cold/warm parameter, i.e. the strength of the starting color [0,1]
+- `d1`       - depth of the ending color in the first part [0,1]
+- `d2`       - depth of the ending color in the latter part [0,1]
+- `wcolor`   - starting color, i.e. the middle color (warmness)
+- `dcolor1`  - ending color of the first part (depth)
+- `dcolor2`  - ending color of the latter part (depth)
+- `logscale` - `true`/`false` for toggling logspacing
 """
 function diverging_palette(h1,
                            h2,
                            N::Int=100;
                            mid=0.5,
-                           c=0.88,
-                           s=0.6,
-                           b=0.75,
                            w=0.15,
                            d1=0.0,
                            d2=0.0,
-                           wcolor=RGB(1,1,0),
-                           dcolor1=RGB(1,0,0),
-                           dcolor2=RGB(0,0,1),
-                           logscale=false)
+                           c=0.88,
+                           s=0.6,
+                           b=0.75,
+                           wcolor =RGB(1.0, 1.0, 0.0),
+                           dcolor1=RGB(1.0, 0.0, 0.0),
+                           dcolor2=RGB(0.0, 0.0, 1.0),
+                           logscale::Bool=false)
+    F = Float64
+    return _diverging_palette(N, F(mid), logscale,
+                              F(h1), F(h2), F(w), F(d1), F(d2), F(c), F(s), F(b),
+                              RGB{F}(wcolor), RGB{F}(dcolor1), RGB{F}(dcolor2))
+end
+function _diverging_palette(N, mid, logscale, h1, h2, w, d1, d2, c, s, b, wcolor, dcolor1, dcolor2)
+    n = isodd(N) ? (N - 1) : N
+    N1 = max(ceil(Int, mid * n), 1)
+    N2 = Int(max(n - N1, 1))
+
+    pal1 = _sequential_palette(N1 + 1, logscale, h1, w, d1, c, s, b, wcolor, dcolor1)
+    pal2 = _sequential_palette(N2 + 1, logscale, h2, w, d2, c, s, b, wcolor, dcolor2)
 
     if isodd(N)
-        n=N-1
+        midcol = weighted_color_mean(0.5, pal1[1], pal2[1])
+        return [pal1[end:-1:2]; midcol; pal2[2:end]]
     else
-        n=N
-    end
-    N1 = max(ceil(Int, mid*n), 1)
-    N2 = Int(max(n-N1, 1))
-
-    pal1 = sequential_palette(h1, N1+1, w=w, d=d1, c=c, s=s, b=b, wcolor=wcolor, dcolor=dcolor1, logscale=logscale)
-    pal1 = reverse(pal1; dims=1)
-
-    pal2 = sequential_palette(h2, N2+1, w=w, d=d2, c=c, s=s, b=b, wcolor=wcolor, dcolor=dcolor2, logscale=logscale)
-
-    if isodd(N)
-        midcol = weighted_color_mean(0.5, pal1[end], pal2[1])
-        return [pal1[1:end-1]; midcol; pal2[2:end]]
-    else
-        return [pal1[1:end-1]; pal2[2:end]]
+        return [pal1[end:-1:2]; pal2[2:end]]
     end
 end
 
@@ -270,21 +275,21 @@ end
 # Main function to handle different predefined colormaps
 #
 """
-    colormap(cname, N=100; mid=0.5, logscale=false, kvs...])
+    colormap(cname, N=100; logscale=false, <keyword arguments>)
 
 Returns a predefined sequential or diverging colormap computed using
 the algorithm by Wijffelaars, M., et al. (2008).
 
 Sequential colormaps `cname` choices are:
 
-- `Blues`
-- `Greens`
-- `Grays`
-- `Oranges`
-- `Purples`,
-- `Reds`
+- `"Blues"`
+- `"Grays"`
+- `"Greens"`
+- `"Oranges"`
+- `"Purples"`
+- `"Reds"`
 
-Diverging colormap choices are `RdBu`.
+Diverging colormap choices are `"RdBu"`.
 
 Optionally, you can specify the number of colors `N` (default 100).
 
@@ -294,49 +299,33 @@ Extra control is provided by keyword arguments.
 You can also use keyword argument names that match the argument names in
 [`sequential_palette`](@ref) or [`diverging_palette`](@ref).
 """
-function colormap(cname::String, N::Int=100; mid=0.5, logscale::Bool=false, kvs...)
-
-    cname = lowercase(cname)
-    if haskey(colormaps_sequential, cname)
-        allowedkeys = [:h, :w, :d, :c, :s, :b, :wcolor, :dcolor]
-        p = copy(colormaps_sequential[cname][1:8])
-
-        for (k,v) in kvs
-            ind = findfirst(e->e==k, allowedkeys)
-            ind == nothing && throw(ArgumentError("Unknown keyword argument $k"))
-            if ind > 0
-                p[ind] = v
-            end
-        end
-
-        # To avoid invalidation risk, it's best to make the call in a manner that inference knows the types
-        return sequential_palette(f64(p[1]), N, w=f64(p[2]), d=f64(p[3]), c=f64(p[4]),
-                                  s=f64(p[5]), b=f64(p[6]), wcolor=rgb8(p[7]),
-                                  dcolor=rgb8(p[8]), logscale=logscale)
-
-    elseif haskey(colormaps_diverging, cname)
-        allowedkeys = [:h1, :h2, :w, :d1, :d2, :c, :s, :b, :wcolor, :dcolor1, :dcolor2]
-        p = copy(colormaps_diverging[cname][1:11])
-
-        for (k,v) in kvs
-            ind = findfirst(e->e==k, allowedkeys)
-            ind == nothing && throw(ArgumentError("Unknown keyword argument $k"))
-            if ind > 0
-                p[ind] = v
-            end
-        end
-
-        return diverging_palette(f64(p[1]), f64(p[2]), N, w=f64(p[3]), d1=f64(p[4]), d2=f64(p[5]),
-                                 c=f64(p[6]), s=f64(p[7]), b=f64(p[8]), wcolor=rgb8(p[9]),
-                                 dcolor1=rgb8(p[10]), dcolor2=rgb8(p[11]), mid=mid,
-                                 logscale=logscale)
-
-    else
-        throw(ArgumentError(string("Unknown colormap: ", cname)))
-    end
+function colormap(cname::AbstractString, N::Integer=100; kvs...)
+    _colormap(String(cname), Int(N), kvs)
 end
-
-colormap(cname::AbstractString, args...; kwargs...) = colormap(String(cname), args...; kwargs...)
-
-f64(x) = Float64(x)::Float64
-rgb8(c) = RGB{N0f8}(c)::RGB{N0f8}
+function _colormap(cname::String, N::Int, kvs)
+    logscale = get(kvs, :logscale, false)::Bool
+    lcname = lowercase(cname)
+    if haskey(colormaps_sequential, lcname)
+        pbs = colormaps_sequential[lcname]
+        keys_s = (:h, :w, :d, :c, :s, :b, :wcolor, :dcolor)
+        for k in keys(kvs)
+            k === :logscale && continue
+            k in keys_s || throw(ArgumentError("Unknown keyword argument: $k"))
+        end
+        ps(i) = oftype(pbs[i], get(kvs, keys_s[i], pbs[i]))
+        return _sequential_palette(N, logscale, (ps(i) for i in eachindex(pbs))...)
+    end
+    if haskey(colormaps_diverging, lcname)
+        pbd = colormaps_diverging[lcname]
+        keys_d = (:h1, :h2, :w, :d1, :d2, :c, :s, :b, :wcolor, :dcolor1, :dcolor2)
+        for k in keys(kvs)
+            k === :logscale && continue
+            k === :mid && continue
+            k in keys_d || throw(ArgumentError("Unknown keyword argument: $k"))
+        end
+        mid = Float64(get(kvs, :mid, 0.5))
+        pd(i) = oftype(pbd[i], get(kvs, keys_d[i], pbd[i]))
+        return _diverging_palette(N, mid, logscale, (pd(i) for i in eachindex(pbd))...)
+    end
+    throw(ArgumentError(string("Unknown colormap: ", cname)))
+end