… src/conversions.jl
… src/maps_data.jl
… src/utilities.jl
… test/colormaps.jl