@kimikage 2021-08-31
Remaining problems:
- Precompilation of `RGB`-->`LCH` conversion (PR #511 can fix this)
- Over-specialization with keyword arguments
- Precompilation of `vcat` in `_diverging_palette`

These can be tackled in separate PRs, though.